Fix data loss editing card
The problem we were seeing was that, after updating a card and immediately clicking back into it, you would see the old data, resulting in data loss.

The first cause of this was priming the card from what was in the list. In cases where it took the list a few seconds to update, this would result in the old data being used.

Even after fixing this, though, we were still served stale card data from the cache. To fix this, we remove the card from the cache entirely upon click, to ensure we will get the latest data from the server before showing the form.